+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * pool: add `RelayPool::force_remove_relay` method ([Yuki Kishimoto])
 * pool: add `RelayFiltering::overwrite_public_keys` method ([Yuki Kishimoto])
 * zapper: add `WebLNZapper` struct (moved from `nostr-webln` crate) ([Yuki Kishimoto])
+* ffi(nostr): add `tag_kind_to_string` func ([Yuki Kishimoto])
 * book: add some examples ([RydalWater])
 
 ### Fixed